[1.1.2#1]Componentize AppTabsScreen & Change splashscreen to recommended code
