+++ App.tsx
@@ -1,95 +1,37 @@
 import * as React from 'react';
-import { AppLoading,SplashScreen  } from 'expo';
-import { KeyboardAvoidingView,SafeAreaView ,View,Image,Platform } from 'react-native';
-import { MaterialCommunityIcons, } from '@expo/vector-icons';
+import { AppLoading } from 'expo';
+import * as SplashScreen from 'expo-splash-screen';
+import {
+  KeyboardAvoidingView,
+  SafeAreaView,
+  View,
+  Image,
+  Platform,
+} from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { useFonts, Inter_900Black } from '@expo-google-fonts/inter';
 
-import HomeScreen from './screens/HomeScreen'
-import ProfileScreen from './screens/ProfileScreen'
-import TipsScreen from './screens/TipsScreen'
+import AppTabsScreen from './screens/AppTabsScreen';
 import defaultStyles from './styles';
 
-const AppTab = createBottomTabNavigator();
-
-const AppTabsScreen = () => (
-  <AppTab.Navigator
-    initialRouteName="Home"
-    tabBarOptions={{
-      activeTintColor: '#d9376e',
-      inactiveTintColor: '#fedbd0',
-      labelStyle: {
-        fontSize: 12,
-        fontFamily:'Inter_900Black'
-      },
-      style: {
-        backgroundColor:'#442c2e',
-        height:50,
-      },
-    }}
-  >
-    <AppTab.Screen
-      name="Home"
-      component={HomeScreen}
-      options= {{
-        tabBarIcon: props => (
-          <MaterialCommunityIcons
-            name={ props.focused ? 'home' : 'home-outline' }
-            size={props.size}
-            color={props.color}
-          />
-        ),
-      }}
-    />
-    <AppTab.Screen
-      name="Profile"
-      component={ProfileScreen}
-      options= {{
-        tabBarIcon: props => (
-          <MaterialCommunityIcons
-            name={ props.focused ? 'heart' : 'heart-outline' }
-            size={props.size}
-            color={props.color}
-          />
-        )
-      }}
-    />
-    <AppTab.Screen
-      name="Tips"
-      component={TipsScreen}
-      options= {{
-        tabBarIcon: props => (
-          <MaterialCommunityIcons
-            name={ props.focused ? 'file' : 'file-outline' }
-            size={props.size}
-            color={props.color}
-          />
-        )
-      }}
-    />
-  </AppTab.Navigator>
-)
-
 const App = () => {
-
-  const [isReady,setReady] = React.useState(false);
+  const [isReady, setReady] = React.useState(false);
 
   const updateAsync = async () => {
     await sleep(2000);
-    SplashScreen.hide();
+    SplashScreen.hideAsync();
     setReady(true);
-  }
-  
+  };
+
   const sleep = (sec) => {
-    return new Promise(resolve => {
-        setTimeout(resolve, sec);
-    })
-  }
-  
-  React.useEffect(()=>{
-    SplashScreen.preventAutoHide();
-  },[]);
+    return new Promise((resolve) => {
+      setTimeout(resolve, sec);
+    });
+  };
+
+  React.useEffect(() => {
+    SplashScreen.preventAutoHideAsync();
+  }, []);
 
   let [fontsLoaded] = useFonts({
     Inter_900Black,
@@ -99,10 +41,10 @@ const App = () => {
     return (
       <View style={[defaultStyles.container]}>
         <Image
-          resizeMode='contain'
-          style={{width: "100%"}}
-          source={require('./assets/splash.png')} 
-          onLoad={updateAsync} 
+          resizeMode="contain"
+          style={{ width: '100%' }}
+          source={require('./assets/splash.png')}
+          onLoad={updateAsync}
         />
       </View>
     );
@@ -113,12 +55,15 @@ const App = () => {
   } else {
     return (
       <NavigationContainer>
-        <KeyboardAvoidingView style={{ flex: 1 }} behavior={Platform.OS == 'ios' ? 'padding' : 'height'}>
+        <KeyboardAvoidingView
+          style={{ flex: 1 }}
+          behavior={Platform.OS == 'ios' ? 'padding' : 'height'}
+        >
           <AppTabsScreen />
-          <SafeAreaView style={{backgroundColor:"#442c2e"}}/>
+          <SafeAreaView style={{ backgroundColor: '#442c2e' }} />
         </KeyboardAvoidingView>
       </NavigationContainer>
-    )
+    );
   }
 };
