+++ app.json
@@ -2,7 +2,7 @@
   "expo": {
     "name": "Pelleculator",
     "slug": "pelleculator",
-    "version": "1.1.1",
+    "version": "1.1.2",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "splash": {
@@ -20,7 +20,7 @@
       },
       "supportsTablet": true,
       "bundleIdentifier": "com.totocalcio.pelleculator",
-      "buildNumber": "1.1.1"
+      "buildNumber": "1.1.2"
     },
     "android": {
       "package": "com.totocalcio.pelleculator",
@@ -28,7 +28,7 @@
         "backgroundColor": "#fedbd0",
         "foregroundImage": "./assets/adaptiveicon.png"
       },
-      "versionCode": 6
+      "versionCode": 7
     },
     "notification": {
       "icon": "./assets/notificationicon.png",
